Refactors upload types for better modularity
